Revert "Sources needed for API moved to commons module."
This reverts commit 1ce8635e55e57072baadebf87a2f5c5e53d14532.
