+++ frontend/src/main/java/cz/cuni/mff/xrg/odcs/frontend/dpu/validator/DataUnitNamesValidator.java
@@ -1,16 +1,17 @@
-package cz.cuni.mff.xrg.odcs.commons.app.dpu.validator;
+package cz.cuni.mff.xrg.odcs.frontend.dpu.validator;
+
+import java.util.HashSet;
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 import cz.cuni.mff.xrg.odcs.commons.app.data.DataUnitDescription;
 import cz.cuni.mff.xrg.odcs.commons.app.dpu.DPUExplorer;
 import cz.cuni.mff.xrg.odcs.commons.app.dpu.DPUTemplateRecord;
 import cz.cuni.mff.xrg.odcs.commons.app.module.DPUValidator;
 import cz.cuni.mff.xrg.odcs.commons.app.module.DPUValidatorException;
-import cz.cuni.mff.xrg.odcs.commons.app.i18n.Messages;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-import java.util.HashSet;
-import java.util.List;
+import cz.cuni.mff.xrg.odcs.frontend.i18n.Messages;
 
 /**
  * Check for duplicity in names of input and output data units.
@@ -34,7 +35,7 @@ class DataUnitNamesValidator implements DPUValidator {
      * exception.
      * 
      * @param dataUnits
-     * @throws cz.cuni.mff.xrg.odcs.commons.app.module.DPUValidatorException
+     * @throws DPUValidatorException
      */
     private void check(List<DataUnitDescription> dataUnits) throws DPUValidatorException {
         HashSet<String> names = new HashSet<>();
