+++ frontend/src/main/java/cz/cuni/mff/xrg/odcs/frontend/dpu/wrap/DPUTemplateWrap.java
@@ -1,12 +1,11 @@
-package cz.cuni.mff.xrg.odcs.commons.app.dpu.wrap;
-
-import cz.cuni.mff.xrg.odcs.commons.app.dpu.DPUTemplateRecord;
-import cz.cuni.mff.xrg.odcs.commons.app.facade.ModuleFacade;
+package cz.cuni.mff.xrg.odcs.frontend.dpu.wrap;
 
 import java.util.Locale;
 
+import cz.cuni.mff.xrg.odcs.commons.app.dpu.DPUTemplateRecord;
+
 /**
- * Wrap {@link cz.cuni.mff.xrg.odcs.commons.app.dpu.DPUTemplateRecord} to made work with configuration and
+ * Wrap {@link DPUTemplateRecord} to made work with configuration and
  * configuration dialog easier.
  *
  * @author Petyr
@@ -23,8 +22,8 @@ public class DPUTemplateWrap extends DPURecordWrap {
      *
      * @param dpuTemplate
      */
-    public DPUTemplateWrap(DPUTemplateRecord dpuTemplate, Locale locale, ModuleFacade moduleFacade) {
-        super(dpuTemplate, true, locale, moduleFacade);
+    public DPUTemplateWrap(DPUTemplateRecord dpuTemplate, Locale locale) {
+        super(dpuTemplate, true, locale);
         this.dpuTemplate = dpuTemplate;
     }
